Update StudentGroup.java
I`ve added changes to StudentGroup class.
Now, if the group doesn`t have a student that is appointed head of the class - this student will be added to the group and appointed head of the class.
Also, I`ve changed completedTasksMap type to String, List<Student>.
Also, I`ve added handling of a case, when someone tries to remove head of the class from group. It`s decided to make removal of head of the class impossible. Group will have to choose a different head of the class.